MHD Filtering Wrapper API

ImageDataIO.filter_mhd_gpu — Function

Runs total variation filtering on a set of images.

Arguments

param_path::Dict: Dictionary of locations of data

path_dir_mhd::String: Location of MHD files to filter

t_range: Time points to filter

list_ch: Channels to filter

f_basename::Function: Function that returns MHD filename given time point and channel

mhd_filt_dir_key::String (optional): Key in param_path that maps to the location to store the output MHD files. Default path_dir_mhd_filt

mip_filt_dir_key::String (optional): Key in param_path that maps to the location to store the output MIP files. Default path_dir_MIP_filt

vmax: Contrast setting for png files